refactor(rules): add canonical rule descriptors and remove domain fs leak (#43)
Introduce a single source of truth for built-in rules so adding a rule no
longer requires edits in ~7 scattered places (construction sites, default
enabled flags, per-id metadata, and a hardcoded class-name Set).

- Add src/rules/registry/ruleDescriptors.ts exporting RULE_DESCRIPTORS
  ({ id, defaultEnabled, metadata, create(config), isEnabled? }) plus
  isRuleEnabled() and getBuiltinRuleIds() helpers.
- Move rule metadata to the rules layer (src/rules/registry/ruleMetadata.ts);
  keep infrastructure/RuleMetadata.ts as a backward-compatible barrel so the
  correct hexagonal direction holds (infrastructure -> rules, not the reverse).
- Derive RuleRegistry.getStats() built-in/custom classification from the
  descriptor list instead of a hardcoded constructor-name Set.
- Remove ContextFile.fromFile() (its readFileSync made the domain touch the
  filesystem). Route the GitHub Action through the FileReader adapter, and have
  the two rules that read auxiliary files (MonorepoHierarchy, ImportResolution)
  construct ContextFile from content they read themselves. Domain stays pure.
